[RENAME] #3 main > mypage로 수정
