+++ app/src/main/res/layout/activity_my_page.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/white"
-    tools:context=".presentation.main.MainActivity">
+    tools:context=".presentation.mypage.MyPageActivity">
 
 
     <ImageView
